Mod: `js_denotest/*.js` moved to `test-util/` `util/` - Also organize test code
