+++ src/tests/js_denotest/util/dbfs.test.js
@@ -11,9 +11,9 @@
 
 import { assertThrows, assertAlmostEquals } from "jsr:@std/assert@1";
 
-import { ONE_FRAME_SAMPLES } from "../../static/AUDIO_PARAM.js";
-import { dbfs_float } from "../../static/util/dbfs.js";
-import { generate_rand_float32array } from "./test-util/rand_f32a.js";
+import { ONE_FRAME_SAMPLES } from "../../../static/AUDIO_PARAM.js";
+import { dbfs_float } from "../../../static/util/dbfs.js";
+import { generate_rand_float32array } from "../test-util/rand_f32a.js";
 
 
 Deno.test(async function true_cases(t) {
